change return type from long to int

+++ src/main/java/ru/serji/kuklyasha/service/DollService.java
@@ -17,5 +17,5 @@ public interface DollService {
 
     List<Doll> getLimitByPage(int page, int limit);
 
-    long totalCount();
+    int totalCount();
 }
